refactor fetcher
